Move controllers and service into associated test classes
