--- htmx-spring-boot/src/test/java/io/github/wimdeblauwe/htmx/spring/boot/mvc/TestService.java
@@ -1,10 +1,0 @@
-package io.github.wimdeblauwe.htmx.spring.boot.mvc;
-
-import org.springframework.stereotype.Service;
-
-@Service
-public class TestService {
-    void doSomething(HtmxRequest details) {
-
-    }
-}
